+++ src/command/server/command/setup.rs
@@ -211,6 +211,7 @@ pub async fn build_registry(
         enable_manifest_redirect: config.global.enable_manifest_redirect,
         max_manifest_size_bytes,
         max_blob_size_bytes: config.global.max_blob_size_bytes(),
+        blob_stream_frame_size: config.global.blob_stream_frame_size_bytes(),
         validate_manifest_references: !config.global.allow_missing_manifest_references,
         global_immutable_tags: config.global.immutable_tags,
         global_immutable_tags_exclusions: config.global.immutable_tags_exclusions.clone(),